Issue 3521 Add Filter To Options Passive Scan Rules
Add generic TableFilterPanel component which accepts a JXTable supported
by additions to Messages.properties.
Added filter components and functionality to PolicyPassiveScanPanel,
changed the passive scan rules table to a JXTable (from JTable).
Updated ManageAddOnsDialog to leverage the new generic component.
Removed now unused cfu keys from Messages.properties

(Based on https://github.com/zaproxy/zaproxy/pull/2341)

Fixes zaproxy/zaproxy#3521
