+++ src/lang/Messages.properties
@@ -666,8 +666,6 @@ To install it either:<ul>\
 cfu.label.cfuonstart = Check for updates on start?
 cfu.label.dir.border   = Add-on Directories
 cfu.label.addons.border = Add-ons
-cfu.label.addons.filter = Filter:
-cfu.label.addons.filter.tooltip = The add-ons filtering system supports regular expressions.
 cfu.label.norecentcfu =  You have not checked for updates for over 3 months.\nZAP is updated regularly, so you are probably\n\
 running with out of date add-ons.\n\nCheck for new updates now?\n\n
 cfu.label.oldzap = This version of ZAP was created over a year ago!\nZAP is updated regularly, so you are probably\n\
@@ -1165,6 +1163,8 @@ footer.scans.label           = Current Scans
 
 form.dialog.button.cancel = Cancel
 
+generic.filter.label = Filter:
+generic.filter.tooltip = The filtering system supports regular expressions.
 generic.options.panel.security.protocols.title = Security Protocols
 generic.options.panel.security.protocols.ssl2hello.label = SSLv2Hello
 generic.options.panel.security.protocols.ssl3.label = SSL 3
